requirements.txt
face-recognition==1.3.0
opencv-python==4.8.1.78
numpy==1.24.3
mysql-connector-python==8.2.0
python-dotenv==1.0.0